02-use-cases/SRE-agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sre-agent"
version = "1.0.0"
description = "Multi-agent SRE assistant for infrastructure troubleshooting"
authors = [{name = "SRE Team", email = "sre@company.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "langgraph>=0.2.39",
    "langchain-core>=0.3.15",
    "langchain-aws>=0.2.6",
    "langchain-anthropic>=0.2.4",
    "langchain-mcp-adapters>=0.1.0",
    "pydantic>=2.0.0",
    "uvloop>=0.20.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pyyaml>=6.0.1",
    "httpx>=0.25.0",
    "click>=8.1.0",
    "mcp>=1.10.1",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.57.1",
    "python-multipart>=0.0.6",
    "boto3==1.39.7",
    "botocore==1.39.7",
    "awscli==1.41.7",
    "bedrock-agentcore==0.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "flake8>=6.0.0",
    "bandit>=1.7.0",
    "pre-commit>=3.0.0",
    "ruff>=0.1.0"
]


[project.scripts]
sre-agent = "sre_agent.cli:main"
sre-gateway-setup = "gateway.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sre_agent*", "gateway*", "backend*"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"